delete listOpt in registerUser.jsp/modifyUser.jsp, ddl, dml add resize image function update follow request routine
