+++ HUBProject_v0.01/WebContent/db/HUB.sql
@@ -22,7 +22,6 @@ CREATE TABLE USER_TB(
 	USERID VARCHAR2(20),
 	PW VARCHAR2(30),
 	EMAIL VARCHAR2(50),
-	LISTOPTION NUMBER,
 	PICTURE VARCHAR2(50),
 	INTRODUCE VARCHAR2(2000),
 	ADMIN VARCHAR2(3)
@@ -123,7 +122,7 @@ CREATE TABLE COCONN_TB(
 	ON DELETE CASCADE
 );
 
-CREATE TABLE USERPOST(
+CREATE TABLE USERPOST_TB(
 	PRIMARY KEY(USERID, POSTID),
 	USERID VARCHAR2(20),
 	POSTID NUMBER,
