+++ HUBProject_v0.01/WebContent/views/list/detailBucketlist.jsp
@@ -79,12 +79,45 @@ h1 {
 		</tr>
 	</table>
 
-	<form action="list/list.do" method="post">
+	<form action="${pageContext.request.contextPath}/list/list.do" method="post">
 		<div class="input-append pull-right">
 			<button class="btn btn-xs btn-default btn-block" type="submit">돌아가기</button>
 		</div>
-		<input type="hidden" name="listOpt" value="${listOpt }">
-		<input type="hidden" name="connChain" value="${connChain }">
+		<input type="hidden" name="listOpt" value="${sessionScope.listOpt }">
+		<input type="hidden" name="connChain" value="${searchWord }">
 	</form>
+	
+<!-- 	<form action="${pageContext.request.contextPath}/follow/request.do" method="post">
+		<div class="input-append pull-right">
+			<button class="btn btn-xs btn-default btn-block" type="submit">팔로우신청</button>
+			<input type="hidden" name="userId" value="${sessionScope.userId }">
+			<input type="hidden" name="followId" value="${bucketlist.userId }">
+			<input type="hidden" name="relation" value="${listOpt }">
+		</div>
+	</form> -->
+	
+	<div class="input-append pull-right">
+		<button class="btn btn-xs btn-default btn-block" id="requestButton" type="button">팔로우요청</button>
+	</div>
+	
+	<script>
+		var userId = "${sessionScope.userId}";
+		var followId = "${bucketlist.userId}";
+		var relation = ${sessionScope.listOpt};
+		
+		$("#requestButton").click(function (){
+			$.ajax({
+				type : 'POST',
+				url : '${pageContext.request.contextPath}/follow/request.do',
+				data : {
+					userId : userId,
+					followId : followId,
+					relation : relation
+				},
+				success : 
+					alert("팔로우를 요쳥했습니다.");
+			});
+		});
+	</script>
 </body>
 </html>
